requirements.txt
PyPDF2==3.0.1
python-docx==1.1.0
sentence-transformers==2.7.0
openai==1.30.0
chromadb==0.5.0
sqlalchemy==2.0.30
pyarabic==0.6.15
streamlit==1.35.0
python-multipart==0.0.9
nltk==3.8.1
python-dotenv==1.0.1
tqdm==4.66.4
loguru==0.7.2
